// src/lib/types.ts
export interface basicInfo {
  name: string
  telNumber: string
  email: string
  sex: string
  birthday: string
  workTime: string
  edcationBack: string
}

export interface work {
  id: string
  company: string
  position: string
  startTime: string
  endTime: string
  responsibility: string
}

export interface project {
  id: string
  name: string
  position: string
  responsibility: string
  description: string
}